Update Player.cpp
Added the telishield key for the player def and script to use.  This key disables the player taking damage and can be checked in the script for stuff like changing skins, etc.

cross hair fix:  pm_modelview .  This cvar changes the origin of the playerview to a joint for the crosshairs.  This is set to a joint close to the player weapons to prevent the trace line used for cross hairs stopping too early when shooting around corners. 
Additionally pm_modelview's axis is now the renderview axis and not the joint other wise the thirdperson crosshair is displayed in wrong directions.
pm_modelview is now on by default.