Replaced stamina system with shield system
Stamina is now used for the player's shield spell.  hold down the old zoom key to activate while the player has stamina.  Calls are made to the player's script to UpdateSkin.  UpdateSkin is new and not apart of vanilla d3 or Rivensin 2010 build
Stamina use also is no longer associated with player's movement and does not adjust movement speed.
The telishield key is used via the player script file to check if active and change the torso animations to blocking/spell casting.