Increased max globals
Needed if you want to play some of the vanilla doom 3 maps by default.  Note that you will need to update more stuff like statements, globals, etc. for the game to start with all of them.  Another note that the scripts are not loaded by default in ruiner/Rivensin due to none of the maps being tested.  Most of them are not third person friendly.  Someday I hope to get a mini version of the campaign ported.