cursor and ingame menu scaling
Cursor now uses 2 different guis based upon aspect ratio just like
the hud. This is due to the translations of of the cross hairs being
off if a 4:3 cursor is on top of a 16:9 render view.
Requires additional key in the player def just like the hud does.
In game menu no longer stretches. Looks much better.
This is the menu that displays the advanced controls.

(DG: based on revility's commit with same message, removed stuff already
 implemented here and cleaned up the rest a bit)