Fix some (mostly Savegame-related) Crashes
Somehow I must have added that
savefile->WriteFloat( spreadCrouchFactor );
twice when merging the Ruiner/Rivensin changes into the SDK.. damn.
(As it's only read once, all data read from the savegame after it
 will be garbage)

I had a crash when creating a Savegame in idAnimState::Save(),
because animator was NULL - so I added a check for that.

The rest has been found with GCC address sanitizer (ASan)