Merged base and d3xp together.
After this commit, all of the DOOM 3 base code and Resurrection of Evil
SDK code exist as one unified codebase, producing two DLLs but sharing
90% of the same code. Anything that is different between them (should)
be `#ifdef`'d out.

The primary motivation for doing this is to eliminate code reuse. I want
to develop a mod for both Resurrection of Evil and the base game, but I
don't want to make my changes twice. Thus, having a unified codebase
like this makes a lot of sense.